+++ cachix/src/Cachix/Client/Exception.hs
@@ -23,6 +23,8 @@ data CachixException
   | ImportUnsupportedHash Text
   | RemoveCacheUnsupported Text
   | InvalidStorePath Text
+  | NixConfParseError Text
+  | NixConfWriteFailed Text
   deriving (Show, Typeable)
 
 instance Exception CachixException where
@@ -46,3 +48,5 @@ instance Exception CachixException where
   displayException (ImportUnsupportedHash s) = toS s
   displayException (RemoveCacheUnsupported s) = toS s
   displayException (InvalidStorePath s) = toS s
+  displayException (NixConfParseError s) = toS s
+  displayException (NixConfWriteFailed s) = toS s